refactor test generator and extract testcase class
